add support for iterating & tree-printing a filesystem directory tree
This involved making the DirTree structures aware of whether they're
a root node or not. Because of how WalkDir iterates through paths from
the filesystem, the root node of the tree is always the directory we're
looking at and don't want nested under a duplicate level.

Now we have a very limited colorless version of tree(1) that produces
nearly the same output.