add colored output
lscolors is a pretty good library, but it's missing a couple features.
It integrates with ansi_term, but I really prefer termcolor for various
reasons so there's some extension traits to convert lscolors Style and
Color objects to their termcolor equivalents.

The main thing missing is a way to get the style for an arbitrary
filename that doesn't exist on disk. I opened an issue on lscolors for
that API, and will have to fork and implement that myself.