libarchive: WIP: start work on the safe rust wrappers
Implementing ArchiveReader is trickier than expected because of
libarchive's use of callback functions, which I want to use so that
I can use any io::Read object as an input to an ArchiveReader.

There's a lot of scary unsafe and Pin stuff here, but I think it should
work. Still lots of work to do before any of this can be tested.