refactor pacman support to handle 'provides' aliases
Now we walk the full pacman local db, parse all the desc files and keep
track of the package path, name, and provides list. When finding
a package, search for an exact package name match, and if not found then
check the providers list (which maps back to real package names).

TODO: the current implementation will re-load the entire pacman database
for each package name requested. Need more refactoring so we can keep
the Pacman object around for multiple queries.