refactor package manager handling and input module
No more InputKind, only PineTree, whose fields are now public so that
package manager code can directly construct a PineTree.

Package trait is gone and PackageManager is made object-safe so that we
can return the system package manager as Box<dyn PackageManager>. No
more intermediate Package type, instead PackageManager::read_package
directly produces the PineTree.

The external behavior is mostly the same here, but now we don't reload
the package manager for every requested package, instead the code in
main handles saving a PackageManager for repeated queries.

One notable external change is that we no longer bail on the first
error. Instead, if there's problems processing one input, that is noted,
and we keep going with the rest of the inputs.