libarchive: add a build.rs hack to recognize a custom libarchive.so symlink
Sometimes libarchive is installed but libarchive-dev isn't, and I don't
have root on the server and don't want to futz with homedir-installing
it myself (especially true since -dev packages for the various
libarchive compression libraries might be missing).

With this, if pkg-config can't fine libarchive, then look for
a libarchive.so symlink in either pine/ or pine/libarchive/ that the
user would create manually. Link against that libarchive.so as a way of
pretending that the -dev package is really installed.