input: fix root directory handling when reading from filesystem
"pine ." or "pine ../foo" are broken because we're sending the full
input path to the DirTree, which ignores "." path entries and fails on
".." entries. Plus passing an absolute path nests everything several
layers deep in the tree, rather than just starting from the path
specified on the command line.

Fix by getting the absolute canonical path for the input path, stripping
that absolute path prefix before adding to the DirTree, and using the
user input as the virtual root entry to print.