make clippy happy
It used to complain about method calls in unwrap_or(...), not it's
complaining that I don't need a closure.

I'm not sure how this lint works, because yeah .len() is a method call
but it's also a trivial integer read. Maybe clippy knows this because
it's a const fn? Or maybe it's hard-coded as "trivial and easy"?