libarchive: regenerate FFI bindings
Generated with libarchive 3.6.2 and bindgen 0.71.1. The bindgen updates
required basically a rewrite of all the options, changing to
allowlist/blocklist terminology, and updating regex usage. It was a big
pain in the butt to make the case-insensitive regexes work, but
I figured out and documented the reasons.

There's also a new --clang-macro-fallback option which can be used to
remove the hacks I had to do for AE_IF* file type macros, though I then
had to replace it with a different hack to switch them from i32 to u32.